REFACTOR : 컬럼 관리 기능
: 	•	ROLE_MANAGER : 칼럼 관리에 대한 모든 권한을 부여합니다. (생성, 수정, 삭제, 순서 이동)
  	•	ROLE_USER : 칼럼 조회에만 접근할 수 있는 권한을 부여합니다.

UserDetailsImpl : 사용자 역할 할당하도록 수정

+++ src/main/java/com/nbcamp/b4trello/service/ColumnService.java
@@ -6,6 +6,10 @@ import com.nbcamp.b4trello.entity.Column;
 import com.nbcamp.b4trello.repository.BoardRepository;
 import com.nbcamp.b4trello.repository.ColumnRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -22,6 +26,7 @@ public class ColumnService {
     private BoardRepository boardRepository;
 
     public List<Column> getColumns(Long boardId) {
+        checkAuthorization();
         if (!boardRepository.existsById(boardId)) {
             throw new IllegalArgumentException(ErrorMessageEnum.BOARD_NOT_FOUND.getMessage());
         }
@@ -29,6 +34,7 @@ public class ColumnService {
     }
 
     public Column createColumn(Long boardId, String columnTitle) {
+        checkAuthorization();
         if (!boardRepository.existsById(boardId)) {
             throw new IllegalArgumentException(ErrorMessageEnum.BOARD_NOT_FOUND.getMessage());
         }
@@ -46,12 +52,14 @@ public class ColumnService {
     }
 
     public void deleteColumn(Long columnId) {
+        checkAuthorization();
         Column column = columnRepository.findById(columnId)
                 .orElseThrow(() -> new IllegalArgumentException(ErrorMessageEnum.COLUMN_NOT_FOUND.getMessage()));
         columnRepository.delete(column);
     }
 
     public void updateColumnSequence(Long boardId, List<Long> columnIds) {
+        checkAuthorization();
         if (!boardRepository.existsById(boardId)) {
             throw new IllegalArgumentException(ErrorMessageEnum.BOARD_NOT_FOUND.getMessage());
         }
@@ -67,6 +75,13 @@ public class ColumnService {
                     .orElseThrow(() -> new IllegalArgumentException(ErrorMessageEnum.COLUMN_NOT_FOUND.getMessage()));
             column.setColumnSequence(i + 1);
             columnRepository.save(column);
+        }
+    }
+
+    private void checkAuthorization() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_MANAGER"))) {
+            throw new AccessDeniedException("접근이 거부되었습니다");
         }
     }
 }
